requirements.txt
numpy==2.1.1
PyOpenGL==3.1.7
